remove oldest event on backwards pass

+++ src/main/java/org/desilvahendricksoftware/debs/Query1.java
@@ -125,10 +125,10 @@ public class Query1 {
 
 
 		//now we need to feed these features into a window of increasing size. On that window,apply the predict function
-		DataStream<Tuple3<Long, Boolean, Double>> events = features
-				.process(new ProcessFunction<Tuple3<Long,Double,Double>, Tuple3<Long, Boolean, Double>>() {
+		DataStream<Tuple3<Long, Boolean, Integer>> events = features
+				.process(new ProcessFunction<Tuple3<Long,Double,Double>, Tuple3<Long, Boolean, Integer>>() {
 					@Override
-					public void processElement(Tuple3<Long, Double, Double> x_n, Context context, Collector<Tuple3<Long, Boolean, Double>> out) throws Exception {
+					public void processElement(Tuple3<Long, Double, Double> x_n, Context context, Collector<Tuple3<Long, Boolean, Integer>> out) throws Exception {
 						eventDetector.numWindowsProcessedSinceLastEventDetected++;
 						//If an event is not detected and w2 has more than 100 elements, empty the window
 						if (eventDetector.numWindowsProcessedSinceLastEventDetected > 100 && !eventDetector.eventDetected) {
@@ -137,7 +137,7 @@ public class Query1 {
 							eventDetector.numWindowsProcessedSinceLastEventDetected = 0;
 						}
 						w2_builder.add(new Point(x_n.f1, x_n.f2, x_n.f0));
-						Tuple3<Long, Boolean, Double> ret = eventDetector.predict(x_n.f0, w2_builder.toArray(new Point[w2_builder.size()]));
+						Tuple3<Long, Boolean, Integer> ret = eventDetector.predict(x_n.f0, w2_builder.toArray(new Point[w2_builder.size()]));
 						if (ret.f1 == true) {
 							eventDetector.eventDetected = true;
 							eventDetector.numWindowsProcessedSinceLastEventDetected = 0;
