+++ src/main/java/org/desilvahendricksoftware/debs/Result.java
@@ -3,9 +3,9 @@ package org.desilvahendricksoftware.debs;
 public class Result {
     long s;
     int d;
-    double event_s;
+    int event_s;
 
-    public Result(long s, boolean d, double event_s) {
+    public Result(long s, boolean d, int event_s) {
         this.s = s;
         this.d = d == true ? 1 : 0 ;
         this.event_s = event_s;
